Clean-up
- Clean-up
- Moved Prevent_deref to CGAL/iterator.h
