+++ BGL/include/CGAL/boost/graph/graph_traits_HalfedgeDS.h
@@ -36,6 +36,7 @@
 #include <CGAL/basic.h>
 #include <CGAL/boost/graph/iterator.h>
 #include <CGAL/Handle_hash_function.h>
+#include <CGAL/iterator.h>
 
 #ifndef CGAL_NO_DEPRECATED_CODE
 #include <CGAL/boost/graph/halfedge_graph_traits.h>
@@ -44,29 +45,6 @@
 namespace CGAL {
 
 namespace internal {
-
-template<typename I>
-class Prevent_deref
-  : public boost::iterator_adaptor<
-  Prevent_deref<I>
-  , I // base
-  , I // value
-  >
-{
-public:
-  typedef boost::iterator_adaptor<
-  Prevent_deref<I>
-  , I // base
-  , I // value
-  > Base;
-  //  typedef typename Prevent_deref::iterator_adaptor_::reference reference;
-  typedef typename Base::reference reference;
-  Prevent_deref() : Base() {};
-  Prevent_deref(const I& i) : Base(i) {};
-private:
-  friend class boost::iterator_core_access;
-  reference dereference() const { return const_cast<typename boost::remove_reference<reference>::type&>(this->base_reference()); }
-};
 
 // a HDS_halfedge pretending to be an Edge
 template<typename Halfedge_handle>
